Refine banner to use custom icons
